Only define parse_deployment_properties once
