+++ brew/rules.bzl
@@ -34,13 +34,15 @@ def _deploy_brew_impl(ctx):
     files = [
         ctx.file.deployment_properties,
         ctx.file.formula,
-        ctx.file.version_file
+        ctx.file.version_file,
+        ctx.file._common_py
     ]
 
     symlinks = {
         'deployment.properties': ctx.file.deployment_properties,
         'formula': ctx.file.formula,
-        'VERSION': ctx.file.version_file
+        'VERSION': ctx.file.version_file,
+        'common.py': ctx.file._common_py,
     }
 
     if ctx.file.checksum:
@@ -89,6 +91,10 @@ deploy_brew = rule(
             allow_single_file = True,
             default = "//brew/templates:deploy.py"
         ),
+        "_common_py": attr.label(
+            allow_single_file = True,
+            default = "//common:common.py"
+        )
     },
     executable = True,
     outputs = {
