+++ common/BUILD
@@ -17,7 +17,7 @@
 # under the License.
 #
 
-exports_files(["archiver.py"])
+exports_files(["archiver.py", "common.py"])
 
 load("@bazel_skylib//:bzl_library.bzl", "bzl_library")
 
@@ -48,7 +48,8 @@ py_binary(
 
 py_library(
     name = "common",
-    srcs = ["common.py"]
+    srcs = ["common.py"],
+    visibility = ["//visibility:public"]
 )
 
 py_binary(
