+++ maven/BUILD
@@ -24,3 +24,14 @@ py_binary(
     srcs = ["assemble.py"],
     visibility = ["//visibility:public"]
 )
+
+py_binary(
+    name = "deployment_rules_builder",
+    srcs = [
+        "deployment_rules_builder.py"
+    ],
+    deps = [
+        "//common:common"
+    ],
+    visibility = ["//visibility:public"]
+)
